http: add additional completion callbacks
Add completion callbacks for url, status, header_field, & header_value.  If
the parser receives a chunk in the middle of one of these, then we will see
the on_* callback multiple times.  Currently, we must either pass partial
values to another parser OR aggregate values and on transition do something.
The latter leads to code that looks for the transition from the old state in
the new state.

Instead, let llhttp indicate when the transition happens and put the
completion code there.  To that end, this commit adds

	on_url_complete
	on_status_complete
	on_header_field_complete
	on_header_value_complete

The URL completion actually happens after the Protocol/Major/Minor are
determined, but this seems reasonable.  The ensures that the Protocol+Verb
is valid and allows header callbacks to ignore fields based on the
Protocol+Version+Verb+URL.

Benchmark results indicate no change in performance.

Re: tests, stubs were added for the completion events but it was not clear
to me how to update several hundred markdown tests without doing it by hand.
If there is a simple way to add these events to existing md tests, please
let me know.

Validated via upstream changes to https://pypi.org/project/llhttp/

Signed-off-by: Derrick Lyndon Pallas <derrick@pallas.us>
PR-URL: https://github.com/nodejs/llhttp/pull/76
Reviewed-By: Fedor Indutny <fedor@indutny.com>