http: fix header completion callbacks for special fields & empty values
In 4cc13c0, the `on_header_field_complete` callback was added.  However,
special header fields take a different path out than general header fields,
so this callback would not be called for them.  This change hooks that path
up to the callback.  In general, any call to `span.headerField.end()` should
be followed by `onHeaderFieldComplete`.

I preëmptively checked to make sure that there was not a similar mistake in
`on_header_value_complete` and there was, specifically for any field except
Content-Length with a completely empty value.  Empty Content-Length is an
error, and the path that checks this was not hooked up.  Now it is too.  In
general, the header value states should never directly transition back to
`header_field_start`, they should poke `onHeaderValueComplete`.

Issue: nodejs/llhttp#79