api: introduce `llhttp_set_lenient_keep_alive`
Extra request/response after `Connection: close` should be either
discarded (in loose mode), or produce error (in strict mode). However,
to support legacy client we introduce:

* `llhttp_set_lenient_keep_alive`

API call to process the extra request normally. Note that it opens the
server to potential cache poisoning attacks and is dangerous to
enable in the most of situations.