api: limit `HTTP_METHOD_MAP` to HTTP methods
RTSP methods shouldn't mix up with HTTP methods in the method map since
they are supported only with the RTSP protocol. Node.js uses
`HTTP_METHOD_MAP` to generate `http.METHODS` array and this breaks tests
in `body-parser` because it assumes that all of these methods are for
HTTP requests.

PR-URL: https://github.com/nodejs/llhttp/pull/103
Reviewed-By: James M Snell <jasnell@gmail.com>
Reviewed-By: Daniele Belardi <dwon.dnl@gmail.com>
Reviewed-By: Matteo Collina <matteo.collina@gmail.com>
Reviewed-By: Beth Griggs <bgriggs@redhat.com>