api: limit `HTTP_METHOD_MAP` to HTTP methods
RTSP methods shouldn't mix up with HTTP methods in the method map since
they are supported only with the RTSP protocol. Node.js uses
`HTTP_METHOD_MAP` to generate `http.METHODS` array and this breaks tests
in `body-parser` because it assumes that all of these methods are for
HTTP requests.