api: allow returning -1 in spans, doc HPE_USER
`HPE_USER` return value was allowed in span callbacks (`on_body`, etc),
but wasn't in the documentation. Additionally, looking at the return
values of other documented callbacks it was easy to assume that `-1` was
a valid return value for spans as well. Instead of returning it as it
is change it to `HPE_USER` and add automatic error reason with the
span's name.