c-headers: do not list PRI as a valid HTTP method
PRI gets special treatment (parse preamble and error) and even though it
is parsed and allowed after `HTTP/1.x` - it is not technically a method
that could appear in the request parsed with llhttp. Remove the method
from the list of HTTP methods in C headers to prevent confusion and
breakage of `body-parser`.

PR-URL: https://github.com/nodejs/llhttp/pull/105
Reviewed-By: Beth Griggs <bgriggs@redhat.com>
Reviewed-By: Matteo Collina <matteo.collina@gmail.com>